pyproject.toml
[project]
name = "prompt-lib-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "pre-commit>=4.2.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.1",
]


[tool.ruff]
line-length = 88
